feat: lifecycle 설정
