+++ apps/mailer/tsconfig.app.json
@@ -5,7 +5,7 @@
     "module": "commonjs",
     "types": ["node"],
     "emitDecoratorMetadata": true,
-    "target": "es2015"
+    "target": "es2015",
   },
   "exclude": ["**/*.spec.ts"],
   "include": ["**/*.ts"]
